shutdown improvements
This commit maintains a Set of all active sessions. This allows
tryShutdown() to gracefully stop the server properly (as
recommended in the Node HTTP2 documentation). The same Set of
sessions also allows forceShutdown() to be implemented.

This commit also ensures that the two shutdown methods are
idempotent, and that forceShutdown() triggers any outstanding
tryShutdown() callbacks.