suppress stream 'end' events on error
If an error occurs on the server, an error response is sent
back to the client. However, an 'end' event is triggered on the
HTTP2 stream, which in turn, triggers an 'end' event on the
gRPC handler stream. The latter 'end' event should not happen.
This commit prevents it by detecting the destroyed HTTP2 stream
condition.