check for closed sessions in tryShutdown()
Server.prototype.tryShutdown() gracefully closes all of the
HTTP2 sessions still active in the server. This commit updates
that logic to only try to close sessions that haven't had
their 'closed' property set to true. I don't think this is
necessary, but it's better to be safe at the expense of one
line of code coverage.